requirements.txt
fastapi==0.110.1 
uvicorn==0.29.0 
python-multipart==0.0.9
pillow==10.0.1 
pytesseract==0.3.10 
PyMuPDF==1.23.17 
python-dotenv==1.0.1 
scikit-learn==1.4.2 
pandas==2.2.2 
numpy==1.26.4 
matplotlib==3.8.4 
seaborn==0.13.2 
joblib==1.4.2 
google-generativeai==0.5.4